uberf-6576: move migration to new phase Signed-off-by: Alexey Zinoviev <alexey.zinoviev@xored.com>
