+++ models/task/package.json
@@ -40,6 +40,7 @@
     "@hcengineering/model-guest": "^0.6.0",
     "@hcengineering/model-view": "^0.6.0",
     "@hcengineering/model-workbench": "^0.6.1",
+    "@hcengineering/model-activity": "^0.6.0",
     "@hcengineering/platform": "^0.6.9",
     "@hcengineering/tags": "^0.6.12",
     "@hcengineering/task": "^0.6.13",
@@ -47,6 +48,7 @@
     "@hcengineering/ui": "^0.6.11",
     "@hcengineering/view": "^0.6.9",
     "@hcengineering/setting": "^0.6.11",
-    "@hcengineering/notification": "^0.6.16"
+    "@hcengineering/notification": "^0.6.16",
+    "@hcengineering/activity": "^0.6.0"
   }
 }
